recoral/requirements.txt
Flask>=3.0.0
APScheduler>=3.10.4
requests>=2.31.0
python-dateutil>=2.8.0
python-dotenv
instaloader>=4.10
maigret
